finishin task

+++ src/api/routes/productos.routes.js
@@ -7,6 +7,8 @@ const routerProductos = express.Router();
 const container = new Container("./src/api/db/productos.txt");
 app.use(express.static(path.join(__dirname, "./src/api/public")));
 
+let isAdmin = true;
+
 routerProductos.get("/:id", (req, res, next) => {
   try {
     res.send(container.getById(parseInt(req.params.id)));
@@ -15,33 +17,66 @@ routerProductos.get("/:id", (req, res, next) => {
   }
 });
 
-routerProductos.post("/", (req, res, next) => {
-  let data = req.body;
-  data = {
-    title: data.title,
-    price: parseInt(data.price),
-    thumbnail: data.thumbnail,
-  };
-  container.save(data);
+routerProductos.get("/", (req, res, next) => {
+  try {
+    res.send(container.getAll());
+  } catch (error) {
+    res.send({ error: "No existen objetos" });
+  }
+});
 
-  res.redirect("/");
+routerProductos.post("/", (req, res, next) => {
+  if (req.headers.authorization !== "Bearer admin") {
+    isAdmin = false;
+    res.json({
+      error: "-1",
+      descripcion: `ruta ${req.originalUrl} metodo ${req.method} no autorizada`,
+    });
+  } else {
+    const time = new Date();
+    let data = req.body;
+    data = {
+      timestamp: time.toLocaleDateString() + " " + time.toLocaleTimeString(),
+      title: data.title,
+      description: data.description,
+      code: data.code,
+      thumbnail: data.thumbnail,
+      price: parseInt(data.price),
+      stock: data.stock,
+    };
+    container.save(data);
+
+    res.send(data);
+  }
 });
 
 routerProductos.put("/:id", (req, res, next) => {
   try {
-    let data = req.body;
+    if (req.headers.authorization !== "Bearer admin") {
+      isAdmin = false;
+      res.send({
+        error: "-1",
+        descripcion: `ruta ${req.originalUrl} metodo ${req.method} no autorizada`,
+      });
+    } else {
+      const time = new Date();
+      let data = req.body;
 
-    let modifiedObj = {
-      id: parseInt(req.params.id),
-      title: data.title,
-      price: parseInt(data.price),
-      thumbnail: data.thumbnail,
-    };
+      let modifiedObj = {
+        timestamp: time.toLocaleDateString() + " " + time.toLocaleTimeString(),
+        title: data.title,
+        description: data.description,
+        code: data.code,
+        thumbnail: data.thumbnail,
+        price: parseInt(data.price),
+        stock: data.stock,
+      };
 
-    container.deleteById(req.params.id);
-    container.save(modifiedObj);
+      container.deleteById(req.params.id);
+      container.save(modifiedObj);
 
-    res.status(201).send(modifiedObj);
+      res.status(201).send(modifiedObj);
+    }
   } catch (error) {
     res.send({ error: "Objeto no encontrado" });
   }
@@ -49,8 +84,16 @@ routerProductos.put("/:id", (req, res, next) => {
 
 routerProductos.delete("/:id", (req, res, next) => {
   try {
-    container.deleteById(req.params.id);
-    res.send({ message: "Objeto eliminado" });
+    if (req.headers.authorization !== "Bearer admin") {
+      isAdmin = false;
+      res.send({
+        error: "-1",
+        descripcion: `ruta ${req.originalUrl} metodo ${req.method} no autorizada`,
+      });
+    } else {
+      container.deleteById(req.params.id);
+      res.send({ message: "Objeto eliminado" });
+    }
   } catch (error) {
     res.send({ message: "Objeto no encontrado" });
   }
